Switch day 13 to use u64

+++ day-13/src/main.rs
@@ -16,20 +16,20 @@ fn main() {
         }
     }
 
-    let mut t: u128 = 0;
+    let mut t: u64 = 0;
     loop {
         let mut found = true;
         let mut incr = 1;
         for (i, opt_id) in bus_ids.iter().enumerate() {
             if let Some(id) = opt_id {
-                let bus_ts = t + i as u128;
-                if bus_ts % (*id as u128) != 0 {
+                let bus_ts = t + i as u64;
+                if bus_ts % (*id as u64) != 0 {
                     found = false;
                     break;
                 } else {
                     // The next common place will be at least the multiple of all the currently
                     // matching IDs, there aren't any places sooner where the multiples align.
-                    incr *= *id as u128
+                    incr *= *id as u64
                 }
             }
         }
